+++ source/packages/@aws-accelerator/accelerator/lib/stacks/network-stacks/network-prep-stack/resolver-resources.ts
@@ -264,7 +264,7 @@ export class ResolverResources {
         if (vpcItem.vpcRoute53Resolver?.queryLogs) {
           if (this.stack.isManagedByAsea(AseaResourceType.ROUTE_53_QUERY_LOGGING, queryLogName!)) {
             this.stack.addLogs(LogLevel.INFO, `DNS Logging for VPC "${vpcItem.name}" is managed externally`);
-            break;
+            continue;
           }
           if (vpcItem.vpcRoute53Resolver.queryLogs.destinations.includes('s3')) {
             this.stack.addLogs(LogLevel.INFO, `Create DNS query log ${queryLogName}-s3 for central S3 destination`);